requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
PyMuPDF==1.24.10
psycopg2-binary==2.9.9
aiokafka==0.11.0
PyJWT==2.9.0
boto3==1.35.36
botocore==1.35.36
python-multipart==0.0.12
